// src/utils/ui-component-props.types.ts
import { JsonDocsComponent, JSXBase } from '@stencil/core/internal';

export type WebComponentProps = 'attr' | 'css' | 'slots';

// These key values are common through css, slots, and props docs generated by stencil.
export interface DocsBaseEntry {
  name: string;
  docs: string;
}

export type TableGenerator = (
  docs: JsonDocsComponent
) => JSXBase.IntrinsicElements[];